+++ Changelog.md
@@ -30,6 +30,7 @@
 - Fix: Passcodes can't be set in certain situations (e.g. when SecurityCraft is installed in SkyFactory 4)
 - Fix: Server crash when the rift stabilizer is used
 - Fix: A Secure Redstone Interface in receiver mode does not update its neighboring blocks when the signal becomes un-/protected
+- Fix: Jade/WTHIT/TOP don't show the owner of an ownable block that is disguised with another ownable block
 - Misc.: Performance improvements for the Secure Redstone Interface
 
 --------------------------Changelog for v1.9.10.2 of SecurityCraft--------------------------